Stat modifiers save on update of actor and store value of difference in attribute. Added modifiers.js as the ModifierHelpers class to refactor code.
